Move shared to InlineExpressionMappingBuilder
