+++ src/Riok.Mapperly/Descriptors/MappingBodyBuilders/ExpressionMappingBodyBuilder.cs
@@ -1,4 +1,4 @@
-using Microsoft.CodeAnalysis;
+using Riok.Mapperly.Descriptors.MappingBuilders;
 using Riok.Mapperly.Descriptors.Mappings;
 using Riok.Mapperly.Descriptors.Mappings.UserMappings;
 using Riok.Mapperly.Diagnostics;
@@ -13,35 +13,12 @@ public static class ExpressionMappingBodyBuilder
     public static void BuildMappingBody(MappingBuilderContext ctx, UserDefinedExpressionMethodMapping mapping)
     {
         // For Expression mappings, the source and target types come from Expression<Func<TSource, TTarget>>
-        // We need to build an ExpressionMapping using the InlineExpressionMappingBuilderContext
         var sourceType = mapping.ExpressionSourceType;
         var targetType = mapping.ExpressionTargetType;
 
-        var mappingKey = TryBuildMappingKey(ctx, sourceType, targetType);
-        var userMapping = ctx.FindMapping(sourceType, targetType) as IUserMapping;
-        var inlineCtx = new InlineExpressionMappingBuilderContext(ctx, userMapping, mappingKey);
-
-        // Check if there's a user-implemented method that can be inlined (same as IQueryable)
-        if (userMapping is UserImplementedMethodMapping && inlineCtx.FindMapping(sourceType, targetType) is { } inlinedUserMapping)
-        {
-            var expressionMapping = new ExpressionMapping(
-                sourceType,
-                mapping.TargetType, // The return type (Expression<Func<TSource, TTarget>>)
-                inlinedUserMapping,
-                ctx.Configuration.SupportedFeatures.NullableAttributes
-            );
-            mapping.SetDelegateMapping(expressionMapping);
-            return;
-        }
-
-        var delegateMapping = inlineCtx.BuildMapping(mappingKey, MappingBuildingOptions.KeepUserSymbol);
+        var delegateMapping = InlineExpressionMappingBuilder.TryBuildInlineMappingForExpression(ctx, sourceType, targetType);
         if (delegateMapping != null)
         {
-            if (ctx.Configuration.Mapper.UseReferenceHandling)
-            {
-                ctx.ReportDiagnostic(DiagnosticDescriptors.QueryableProjectionMappingsDoNotSupportReferenceHandling);
-            }
-
             var expressionMapping = new ExpressionMapping(
                 sourceType,
                 mapping.TargetType, // The return type (Expression<Func<TSource, TTarget>>)
@@ -53,19 +30,5 @@ public static class ExpressionMappingBodyBuilder
         }
 
         ctx.ReportDiagnostic(DiagnosticDescriptors.CouldNotCreateMapping, sourceType, targetType);
-    }
-
-    private static TypeMappingKey TryBuildMappingKey(MappingBuilderContext ctx, ITypeSymbol sourceType, ITypeSymbol targetType)
-    {
-        // if nullable reference types are disabled
-        // and there was no explicit nullable annotation,
-        // the non-nullable variant is used here.
-        // Otherwise, this would lead to a select like source.Select(x => x == null ? throw ... : new ...)
-        // which is not expected in this case.
-        // see also https://github.com/riok/mapperly/issues/1196
-        sourceType = ctx.SymbolAccessor.NonNullableIfNullableReferenceTypesDisabled(sourceType, ctx.UserMapping?.SourceType);
-        targetType = ctx.SymbolAccessor.NonNullableIfNullableReferenceTypesDisabled(targetType, ctx.UserMapping?.TargetType);
-
-        return new TypeMappingKey(sourceType, targetType);
     }
 }
