+++ src/Riok.Mapperly/Descriptors/MappingBuilders/InlineExpressionMappingBuilder.cs
@@ -1,3 +1,4 @@
+using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Riok.Mapperly.Descriptors.Mappings;
 using Riok.Mapperly.Descriptors.Mappings.UserMappings;
@@ -7,6 +8,41 @@ namespace Riok.Mapperly.Descriptors.MappingBuilders;
 
 public static class InlineExpressionMappingBuilder
 {
+    /// <summary>
+    /// Builds an inline expression delegate mapping for the given source and target types.
+    /// This is used by both queryable projection mappings and expression mappings.
+    /// </summary>
+    /// <param name="ctx">The mapping builder context.</param>
+    /// <param name="sourceType">The source type to map from.</param>
+    /// <param name="targetType">The target type to map to.</param>
+    /// <returns>The built mapping, or <see langword="null"/> if no mapping could be created.</returns>
+    public static INewInstanceMapping? TryBuildInlineMappingForExpression(
+        MappingBuilderContext ctx,
+        ITypeSymbol sourceType,
+        ITypeSymbol targetType
+    )
+    {
+        var mappingKey = BuildMappingKey(ctx, sourceType, targetType);
+        var userMapping = ctx.FindMapping(sourceType, targetType) as IUserMapping;
+        var inlineCtx = new InlineExpressionMappingBuilderContext(ctx, userMapping, mappingKey);
+
+        if (userMapping is UserImplementedMethodMapping && inlineCtx.FindMapping(sourceType, targetType) is { } inlinedUserMapping)
+        {
+            return inlinedUserMapping;
+        }
+
+        var mapping = inlineCtx.BuildMapping(mappingKey, MappingBuildingOptions.KeepUserSymbol);
+        if (mapping == null)
+            return null;
+
+        if (ctx.Configuration.Mapper.UseReferenceHandling)
+        {
+            ctx.ReportDiagnostic(DiagnosticDescriptors.QueryableProjectionMappingsDoNotSupportReferenceHandling);
+        }
+
+        return mapping;
+    }
+
     /// <summary>
     /// Tries to inline a given mapping.
     /// This works as long as the user implemented methods
@@ -79,5 +115,19 @@ public static class InlineExpressionMappingBuilder
 
             _ => null,
         };
+    }
+
+    private static TypeMappingKey BuildMappingKey(MappingBuilderContext ctx, ITypeSymbol sourceType, ITypeSymbol targetType)
+    {
+        // if nullable reference types are disabled
+        // and there was no explicit nullable annotation,
+        // the non-nullable variant is used here.
+        // Otherwise, this would lead to a select like source.Select(x => x == null ? throw ... : new ...)
+        // which is not expected in this case.
+        // see also https://github.com/riok/mapperly/issues/1196
+        sourceType = ctx.SymbolAccessor.NonNullableIfNullableReferenceTypesDisabled(sourceType, ctx.UserMapping?.SourceType);
+        targetType = ctx.SymbolAccessor.NonNullableIfNullableReferenceTypesDisabled(targetType, ctx.UserMapping?.TargetType);
+
+        return new TypeMappingKey(sourceType, targetType);
     }
 }
